Sort posts by last reply. Add last reply timestamp
I have now realised that it would have been a much better design to
simply have merged the index(), index_amount() and post() functions
probably  with a helper function that does most of the work. It has
gotten to the point where there is a bit too much overlap. Everytime I
make a change to one I have to adapt it to each function. Identical
features are being implemented in different ways. Not good :^(